+++ node/resolvers/Mutations/Users.ts
@@ -1,5 +1,5 @@
 import { MessageSFPUserAddError, StatusAddUserError } from '../../constants'
-import GraphQLError from '../../utils/GraphQLError'
+import GraphQLError, { getErrorMessage } from '../../utils/GraphQLError'
 import type { ImpersonateMetricParams } from '../../utils/metrics/impersonate'
 import {
   sendImpersonateB2BUserMetric,
@@ -366,6 +366,59 @@ const Users = {
       sessionData?.namespaces?.account?.accountName,
       sessionData?.namespaces['storefront-permissions']
     )
+  },
+
+  removeUserWithEmail: async (
+    _: void,
+    { orgId, costId, email }: UserArgs,
+    ctx: Context
+  ) => {
+    const {
+      clients: { events, storefrontPermissions: storefrontPermissionsClient },
+      vtex: { logger },
+    } = ctx as any
+
+    storefrontPermissionsClient
+      .getUsersByEmail(email, orgId, costId)
+      .then((result: any) => {
+        const user = result.data.getUsersByEmail[0]
+
+        if (!user) return
+
+        const id = user.id
+        const userId = user.userId
+
+        const fields = {
+          email,
+          id,
+          userId,
+        }
+
+        return storefrontPermissionsClient
+          .deleteUser(fields)
+          .then((response: any) => {
+            events.sendEvent('', 'b2b-organizations-graphql.removeUser', {
+              id,
+              email,
+            })
+            sendRemoveUserMetric(ctx, logger, ctx.vtex.account, fields)
+            return response.data.deleteUser
+          })
+          .catch((error: any) => {
+            logger.error({
+              error,
+              message: 'removeUser-deleteUserError',
+            })
+            return { status: 'error', message: error }
+          })
+      })
+      .catch((error: any) => {
+        logger.error({
+          error,
+          message: 'getUsers-error',
+        })
+        throw new GraphQLError(getErrorMessage(error))
+      })
   },
 
   /**